+++ src/components/GoalList/GoalList.module.css
@@ -1,0 +1,4 @@
+.GoalListTextList {
+    /* 24 / 2 - 7 center of UserPic and center of PlusIcon */
+    margin-left: 5px;
+}
