+++ src/components/GoalPage/GoalPage.tsx
@@ -7,7 +7,7 @@ import { ExternalPageProps } from '../../utils/declareSsrProps';
 import { ModalEvent, dispatchModalEvent } from '../../utils/dispatchModal';
 import { Page } from '../Page/Page';
 import { PageActions } from '../PageActions/PageActions';
-import { PageSep } from '../PageSep';
+import { PageSep } from '../PageSep/PageSep';
 import { IssueKey } from '../IssueKey';
 import { useLocalStorage } from '../../hooks/useLocalStorage';
 import { useWillUnmount } from '../../hooks/useWillUnmount';
