--- src/components/PageSep.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const PageSep = styled.div`
-    margin: 16px 0px;
-
-    box-sizing: border-box;
-
-    border-top: 1px solid var(--gray4);
-`;
