+++ src/components/PageSep/PageSep.module.css
@@ -1,0 +1,5 @@
+.PageSep {
+    margin: var(--gap-m) 0px;
+    box-sizing: border-box;
+    border-top: 1px solid var(--gray4);
+}
