+++ src/components/PageSep/PageSep.tsx
@@ -1,0 +1,3 @@
+import s from './PageSep.module.css';
+
+export const PageSep = () => <div className={s.PageSep} />;
