+++ src/components/Reactions/Reactions.module.css
@@ -1,0 +1,4 @@
+.Reactions {
+    display: flex;
+    gap: var(--gap-xs);
+}
