+++ src/components/UserEditableList/UserEditableList.module.css
@@ -1,0 +1,3 @@
+.UserComboBoxContainer {
+    margin-top: var(--gap-xs);
+}
