+++ src/components/UserEditableList/UserEditableList.tsx
@@ -1,23 +1,16 @@
 import { ComponentProps, FC } from 'react';
-import styled from 'styled-components';
 import { nullable } from '@taskany/bricks';
 import { IconXCircleSolid } from '@taskany/icons';
 import { User } from '@prisma/client';
 
 import { safeUserData } from '../../utils/getUserName';
-import { TextList, TextListItem } from '../TextList';
+import { TextList, TextListItem } from '../TextList/TextList';
 import { UserBadge } from '../UserBadge/UserBadge';
 import { UserComboBox } from '../UserComboBox';
 import { AddInlineTrigger } from '../AddInlineTrigger/AddInlineTrigger';
 
 import { tr } from './UserEditableList.i18n';
-
-const inputHeight = '28px';
-
-const StyledInputContainer = styled.div`
-    margin-top: var(--gap-xs);
-    height: ${inputHeight};
-`;
+import s from './UserEditableList.module.css';
 
 export const UserEditableList: FC<{
     editable?: boolean;
@@ -44,7 +37,7 @@ export const UserEditableList: FC<{
         </TextList>
 
         {nullable(editable, () => (
-            <StyledInputContainer>
+            <div className={s.UserComboBoxContainer}>
                 <UserComboBox
                     placement="bottom-start"
                     placeholder={tr('Type user name or email')}
@@ -52,7 +45,7 @@ export const UserEditableList: FC<{
                     onChange={onAdd}
                     renderTrigger={(props) => <AddInlineTrigger text={triggerText} onClick={props.onClick} />}
                 />
-            </StyledInputContainer>
+            </div>
         ))}
     </div>
 );
